Reduce lock contention in shared caches
There are a number of caches that are shared across the build tree / build session.
These caches are accessed by many threads at once across the build, especially if we are
doing a lot of dependency resolution.
We previously relied on compute and computeIfAbsent only, however these methods rely on
lcoking for atomicity.

Testing on large builds have shown this leads to lock contention and affects performance.
We update the usages of these caches to insteaed first call 'get' on the caches to check
for the value, and only if not present do we use compute and computeIfAbsent.

'get' generally should not perform any locking, unlock computeX, so in most cases after the
value has been calculated for a given key, we should be able to avoid lock contention on
these caches.