Introduce `DynamicGraphCycleDetector`
Initial goal is to prohibit cycles in included builds hierarchy in IP mode. We intentionally don't use `CachingDirectedGraphWalker` because of its nature - it assumed to be used once, when graph has already been constructed. Our need is to detect cycles as we add nodes (included builds in fact) to the graph.